Refactor submodues in extensions
