+++ python/packages/autogen-ext/src/autogen_ext/code_executors/azure/_azure_container_code_executor.py
@@ -48,7 +48,11 @@ class ACADynamicSessionsCodeExecutor(CodeExecutor):
 
     .. note::
 
-        This class requires the :code:`azure` extra for the :code:`autogen-ext` package.
+        This class requires the :code:`azure` extra for the :code:`autogen-ext` package:
+
+        ```bash
+        pip install 'autogen-ext[azure]==0.4.0.dev7'
+        ```
 
 
     **This will execute LLM generated code on an Azure dynamic code container.**
