+++ python/packages/autogen-ext/src/autogen_ext/code_executors/docker/_docker_code_executor.py
@@ -52,7 +52,11 @@ class DockerCommandLineCodeExecutor(CodeExecutor):
 
     .. note::
 
-        This class requires the :code:`docker` extra for the :code:`autogen-ext` package.
+        This class requires the :code:`docker` extra for the :code:`autogen-ext` package:
+
+        .. code-block:: bash
+
+            pip install 'autogen-ext[docker]==0.4.0.dev7'
 
 
     The executor first saves each code block in a file in the working
