+++ python/packages/autogen-ext/src/autogen_ext/tools/langchain/__init__.py
@@ -1,0 +1,3 @@
+from ._langchain_adapter import LangChainToolAdapter
+
+__all__ = ["LangChainToolAdapter"]
