+++ python/packages/autogen-ext/src/autogen_ext/tools/langchain/_langchain_adapter.py
@@ -17,7 +17,11 @@ class LangChainToolAdapter(BaseTool[BaseModel, Any]):
 
     .. note::
 
-        This class requires the :code:`langchain` extra for the :code:`autogen-ext` package.
+        This class requires the :code:`langchain` extra for the :code:`autogen-ext` package:
+
+        .. code-block:: bash
+
+            pip install 'autogen-ext[langchain]'
 
 
     Args:
